requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic-settings>=2.6.0
langchain-google-genai>=2.0.7
langchain-core>=0.3.21
httpx>=0.27.0
python-dotenv>=1.0.0
